% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{check_rise_threshold}
\alias{check_rise_threshold}
\title{Title}
\usage{
check_rise_threshold(profile, threshold = 2.3, min_increase_points = 3)
}
\arguments{
\item{min_increase_points}{}
}
\description{
Title
}
