% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{decimal_to_posixct}
\alias{decimal_to_posixct}
\title{Convert Decimal Hours to POSIXct Timestamps}
\usage{
decimal_to_posixct(decimal_hours, profile_datetime, tz = "UTC")
}
\arguments{
\item{decimal_hours}{A numeric vector representing time in decimal hours.}

\item{profile_datetime}{A vector of POSIXct timestamps where the first element serves as the reference origin.}

\item{tz}{A string specifying the time zone (default is "UTC").}
}
\value{
A vector of POSIXct timestamps.
}
\description{
This function converts a numeric vector of decimal hours back into POSIXct timestamps,
using the first timestamp in \code{profile_datetime} as the origin.
}
