% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_base_segment.R
\name{define_base_segment}
\alias{define_base_segment}
\title{Title}
\usage{
define_base_segment()
}
\arguments{
\item{profile}{a vector of melatonin concentrations}
}
\value{
the segment of the melatonin concentrations defined as belonging to the baseline
}
\description{
Title
}
