% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmoR-package.R
\docType{package}
\name{dlmoR-package}
\alias{dlmoR}
\alias{dlmoR-package}
\title{dlmoR: Determine Time of Dimlight Melatonin Onset}
\description{
Calculate the time of dimlight melatonin onset from a time series of melatonin concentrations, using the hockeystick method.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/salmathalji/dlmoR}
  \item Report bugs at \url{https://github.com/salmathalji/dlmoR/issues}
}

}
\author{
\strong{Maintainer}: Salma Thalji \email{salma.thalji@tuebingen.mpg.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
