% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dlmo.R
\name{.read_melatonin_data}
\alias{.read_melatonin_data}
\title{Helper Function to Read-in Melatonin Data from a CSV-File}
\usage{
.read_melatonin_data(file_path)
}
\arguments{
\item{file_path}{A string specifying the path to the CSV file.}
}
\value{
A data frame with columns \code{datetime} (or \code{time} and \code{melatonin}).
}
\description{
This function reads in a CSV file which ideally contains melatonin concentration data and associated datetime (or time) stamps and returns
a tibble.
}
\keyword{internal}
