% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_parallelogram.R
\name{get_corners}
\alias{get_corners}
\title{Compute the Four Corner Points of a Parallelogram}
\usage{
get_corners(x0, y0, x1, y1, slope)
}
\arguments{
\item{x0}{Numeric. The x-coordinate of the bottom-left corner.}

\item{y0}{Numeric. The y-coordinate of the bottom-left corner.}

\item{x1}{Numeric. The x-coordinate of the bottom-right corner.}

\item{y1}{Numeric. The y-coordinate of the top-right corner.}

\item{slope}{Numeric. The slope of the parallelogram's diagonal edges.}
}
\value{
A list containing the coordinates of the four corners.
}
\description{
Given the bottom-left and top-right coordinates, this function calculates
the four corners of a parallelogram with a specified slope.
}
