% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_parallelogram.R
\name{optimize_parallelogram}
\alias{optimize_parallelogram}
\title{Main Function to Optimize Parallelogram}
\usage{
optimize_parallelogram(x, y)
}
\arguments{
\item{x}{A numeric vector representing x-coordinates (e.g., time in decimal hours).}

\item{y}{A numeric vector representing y-coordinates (e.g., melatonin concentration).}
}
\value{
A list containing optimized parameters for the parallelogram, including corner points.
}
\description{
Optimize Parallelogram to Fit Input Data
}
\details{
This function fits the smallest possible parallelogram that fully encloses all input points.
It minimizes the area while ensuring that all points remain within the defined region.
}
\examples{
optimize_parallelogram(c(18.56, 20.05), c(1.84, 9.203))
}
