% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_parallelogram}
\alias{plot_parallelogram}
\title{Overlay a Parallelogram on a Melatonin Profile Plot}
\usage{
plot_parallelogram(plot, profile_data, pll_result)
}
\arguments{
\item{plot}{A ggplot object. The base plot to which the parallelogram will be added.}

\item{profile_data}{A dataframe containing melatonin concentration data with \code{datetime} values.}

\item{pll_result}{A list containing the \strong{optimized parallelogram parameters}, including:
\itemize{
\item \code{pll_datetime_0}: POSIXct timestamp for the left boundary.
\item \code{pll_datetime_1}: POSIXct timestamp for the right boundary.
\item \code{pll_slope}: Numeric. The slope of the parallelogram's edges.
\item \code{corners}: A list of four corner coordinates (\code{ll}, \code{lr}, \code{ur}, \code{ul}).
}}
}
\value{
A ggplot object with the parallelogram overlay.
}
\description{
This function adds a \strong{parallelogram visualization} to a melatonin profile plot.
The best-fit parallelogram encompasses all ascending points with the smallest area parallelogram
possible. The ratio of parallelogram latera edge to long diagonal is used as a criterion
for trimming the ascending region to ensure that only a sufficiently fast enough melatonin
rise is taken into consideration when fitting the profile and searching for the DLMO point.
}
\details{
\itemize{
\item The \strong{parallelogram} is generated using the \strong{optimized fit} from \code{parallelogram_fit()}.
\item The \strong{edges} of the parallelogram indicate a \strong{bounded region of interest (ROI)}.
\item Two \strong{diagonal dashed lines} highlight the \strong{shape constraints} used during optimization.
\item Uses \strong{red fill} with \strong{transparency (alpha = 0.1)} to avoid obscuring data.
}
}
\examples{
\dontrun{
  plot <- ggplot2::ggplot(profile_data, ggplot2::aes(x = datetime, y = melatonin)) +
      ggplot2::geom_point()
  plot <- plot_parallelogram(plot, profile_data, pll_result)
}
}
