% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_roi}
\alias{plot_roi}
\title{Add Region of Interest (ROI) to a Plot}
\usage{
plot_roi(plot, roi, roi_line_only)
}
\arguments{
\item{plot}{A ggplot object. The base plot to which the ROI should be added.}

\item{roi}{A list containing the \strong{Region of Interest (ROI) coordinates}:
\itemize{
\item \code{x_start} (POSIXct) – The starting time of the ROI.
\item \code{x_end} (POSIXct) – The ending time of the ROI.
\item \code{y_min} (numeric) – The minimum melatonin concentration in the ROI.
\item \code{y_max} (numeric) – The maximum melatonin concentration in the ROI.
}}

\item{roi_line_only}{Logical. If \code{TRUE}, only the \strong{horizontal segment} is plotted.
If \code{FALSE}, a \strong{semi-transparent rectangle} is added to indicate the \strong{full ROI region}.}
}
\value{
A ggplot object with the ROI overlay.
}
\description{
This function overlays the \strong{Region of Interest (ROI)} on an existing melatonin profile plot.
It adds a \strong{horizontal segment} to mark the ROI range and optionally a \strong{shaded rectangle} to highlight the full area that is parsed during the DLMO search.
}
\details{
\itemize{
\item The \strong{horizontal segment} is plotted at \code{y = -0.2} to clearly mark the time range of the ROI.
\item If \code{roi_line_only = FALSE}, the \strong{rectangle spans} from \code{y_min} to \code{y_max} with a light \strong{orchid3} fill.
}
}
\examples{
\dontrun{
  plot <- ggplot2::ggplot(profile_data, ggplot2::aes(x = datetime, y = melatonin)) +
      ggplot2::geom_point()
  roi <- list(x_start = as.POSIXct("2024-04-16 18:00:00"),
              x_end = as.POSIXct("2024-04-16 22:00:00"),
              y_min = 0.2, y_max = 2.5)
  plot_roi(plot, roi, roi_line_only = FALSE)
}
}
