% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{posixct_to_decimal}
\alias{posixct_to_decimal}
\title{Convert POSIXct Timestamps to Decimal Hours}
\usage{
posixct_to_decimal(posix_times, profile_datetime)
}
\arguments{
\item{posix_times}{A vector of POSIXct timestamps.}

\item{profile_datetime}{A vector containing POSIXct timestamps, where the first element serves as the reference origin.}
}
\value{
A numeric vector representing decimal hours since the origin.
}
\description{
This function converts a vector of POSIXct timestamps into decimal hours,
relative to the first timestamp in \code{profile_datetime}.
}
