% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{preprocess_profiles}
\alias{preprocess_profiles}
\title{Title}
\usage{
preprocess_profiles(
  data,
  threshold = 2.3,
  min_increase_points = 3,
  min_points = 3
)
}
\arguments{
\item{min_points}{}
}
\description{
Title
}
