% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dlmo.R
\name{validate_data_structure}
\alias{validate_data_structure}
\title{Function to Validate Data}
\usage{
validate_data_structure(data)
}
\arguments{
\item{data}{A data frame with columns \code{time} and \code{melatonin}.}
}
\value{
Returns a Boolean. If criteria are not met, an error message is returned.
}
\description{
This function validates the structure of an input dataframe. It checks that the dataframe
it contains \code{melatonin} and \code{time} columns and checks that they contain numeric and/or POSIXct data.
}
