% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_ascending_segment.R
\name{define_ascending_segment}
\alias{define_ascending_segment}
\title{Define the ascending segment of a melatonin profile}
\usage{
define_ascending_segment(
  profile_data,
  threshold = 2.3,
  interval_limit = lubridate::hours(2)
)
}
\arguments{
\item{profile_data}{A tibble containing melatonin profile data with the following columns:
\itemize{
\item \code{datetime}: A POSIXct column representing timestamps for each measurement.
\item \code{melatonin}: A numeric column representing melatonin concentrations.
\item \code{base}: A binary column (1 for base segments, 0 otherwise) indicating low melatonin baseline segments.
\item \code{slope}: A numeric column representing the rate of change in melatonin concentrations between consecutive points.
}}

\item{threshold}{Numeric. The melatonin concentration threshold to define the ascending segment (default = 2.3 pg/mL).}

\item{interval_limit}{Numeric or lubridate duration. Specifies the minimum time interval between
consecutive threshold crossings to consider them as independent ascending events. If numeric, it is interpreted as hours (default = 2 hours).}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{datetime}: The original timestamps.
\item \code{melatonin}: The input melatonin concentrations.
\item \code{base}: The input base column (1 for base segments, 0 otherwise).
\item \code{slope}: The calculated slope of the profile.
\item \code{ascending}: A binary indicator (1 for points in the ascending segment, 0 otherwise).
}
}
\description{
This function identifies and labels the ascending segment of a melatonin profile,
where melatonin concentrations rise above a specified threshold and satisfy specific
time interval and slope conditions. It also incorporates additional rules to refine
the identification of ascending points.
}
