% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_base_segment.R
\name{define_base_segment}
\alias{define_base_segment}
\title{Define the base segment of melatonin profile}
\usage{
define_base_segment(profile_tibble, threshold = 2.3)
}
\arguments{
\item{profile_tibble}{A tibble containing datetime, melatonin, and time columns}

\item{threshold}{The melatonin concentration threshold to define the base segment (default = 2.3)}
}
\value{
A tibble with the base segment labeled (1 for base, 0 for non-base), including datetime, time, and melatonin columns
}
\description{
Define the base segment of melatonin profile
}
