% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_df_structure.R
\name{validate_df_structure}
\alias{validate_df_structure}
\title{Function to Validate Data}
\usage{
validate_df_structure(data)
}
\arguments{
\item{data}{A data frame, ideally with columns \code{datetime} (or \code{time}) and \code{melatonin}.}
}
\value{
Returns a dataframe with columns \code{melatonin}, \code{datetime} and \code{time} if the input df contains
\code{melatonin} and \code{datetime} columns. Returns \code{melatonin} and \code{time} columns if the input df only contains these.
If criteria are not met, a corresponding error message is returned.
}
\description{
This function validates the structure of an input dataframe. It verifies that the dataframe
contains \code{melatonin} and \code{datetime} (or \code{time}) columns and checks that they contain numeric and POSIXct (or hms) data, respectively.
}
